broke up analysis module
