+++ autofit/non_linear/analysis/__init__.py
@@ -1,0 +1,1 @@
+from .analysis import Analysis
